Updated to SWIG 1.3.35
"-include fstream" is no longer necessary for cross compilation. It was
needed because "ruby.h" (included from id3lib_api_wrap.cxx) defines the
macros "read" and "write", and the fstream header file (included later
through id3lib) defines methods with the same names. And because macros
are the way they are, those definitions clash. But the newer SWIG
generates code to undef the macros after "ruby.h" is included.